Hot fix

+++ app/Http/Controllers/User/AddressController.php
@@ -27,16 +27,6 @@ class AddressController extends Controller
     }
 
     /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
      * Store a newly created resource in storage.
      *
      * @param \App\Http\Requests\User\AddressCreateRequest $request
@@ -46,7 +36,7 @@ class AddressController extends Controller
     {
         $validated = $request->validated();
         $address = Address::create($validated);
-        $address->users()->attach(auth("sanctum")->user()->id);
+        $address->users()->attach(auth()->user()->id);
         return response(["message" => "Address added successfully"], 201);
     }
 
@@ -58,22 +48,11 @@ class AddressController extends Controller
      */
     public function show($address)
     {
-        $address = Address::where("id", $address)->first();
+        $address = auth()->user()->addresses->where("id", $address)->first();
         if ($address)
             return response(new AddressResource($address));
         return response(["message" => "Address Not Found!"], 404);
 
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param \App\Models\Address $address
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Address $address)
-    {
-        //
     }
 
     /**
@@ -86,7 +65,8 @@ class AddressController extends Controller
     public function update(UpdateRequest $request, Address $address)
     {
         $validated = $request->validated();
-        if ($address->exists) {
+
+        if ($address->exists && auth()->user()->addresses->contains($address)) {
             $address->update($validated);
             return response(["message" => "Address updated successfully"], 201);
         }
@@ -96,12 +76,13 @@ class AddressController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param \App\Models\Address $address
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Address $address)
+    public function destroy(int $id)
     {
-        if ($address->exists) {
+        $address = Address::where("id", $id)->first();
+        if ($address?->exists && auth()->user()->addresses->contains($address)) {
             $address->delete();
             return response(["message" => "Address deleted successfully"]);
         }
@@ -111,12 +92,15 @@ class AddressController extends Controller
 
     public function setDefaultAddress(Address $address)
     {
-        $addresses = auth("sanctum")->user()->addresses;
-        collect($addresses)->map(fn($address) => $address->update(["default" => false]));
-        if ($address->exists) {
+        $addresses = auth()->user()->addresses;
+        if ($addresses->contains($address)) {
+            collect($addresses)->map(function ($userAddress) {
+                return $userAddress->update(["default" => false]);
+            });
+
             $address->update(["default" => true]);
             return response(["message" => "Current address updated successfully"], 201);
         }
-        return response(["message" => "fail to set default address"], 405);
+        return response(["message" => "Fail to set default address! Address not found."], 405);
     }
 }
